refactor: use the new crate
