+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
   "crates/primitives",
   "crates/bridge-proof/protocol",
   "crates/bridge-proof/protocol2",
+  "crates/bridge-proof/primitives",
   "crates/bridge-proof/snark",
   "crates/rpc/bridge",
   "crates/rpc/strata",
@@ -34,6 +35,7 @@ strata-bridge-agent = { path = "crates/agent" }
 strata-bridge-db = { path = "crates/db" }
 strata-bridge-guest-builder = { path = "bridge-guest-builder" }
 strata-bridge-primitives = { path = "crates/primitives" }
+strata-bridge-proof-primitives = { path = "crates/bridge-proof/primitives" }
 strata-bridge-proof-protocol = { path = "crates/bridge-proof/protocol" }
 strata-bridge-proof-protocol2 = { path = "crates/bridge-proof/protocol2" }
 strata-bridge-proof-snark = { path = "crates/bridge-proof/snark" }
