+++ crates/bridge-proof/primitives/Cargo.toml
@@ -13,18 +13,13 @@ rust.unused_must_use = "deny"
 rustdoc.all = "warn"
 
 [dependencies]
-strata-crypto.workspace = true
-strata-l1tx.workspace = true
 strata-primitives.workspace = true
 strata-proofimpl-btc-blockspace.workspace = true
 strata-state.workspace = true
-zkaleido.workspace = true
-zkaleido-native-adapter.workspace = true
 
 arbitrary.workspace = true
 bitcoin.workspace = true
 borsh.workspace = true
-thiserror.workspace = true
 
 [dev-dependencies]
 bincode.workspace = true
