+++ crates/bridge-proof/primitives/src/lib.rs
@@ -1,2 +1,7 @@
-pub mod tx;
-pub mod tx_inclusion_proof;
+//! TODO: Add docs
+
+mod tx;
+
+/// TODO: Add docs
+mod tx_inclusion_proof;
+pub use tx_inclusion_proof::L1TxWithProofBundle;
