+++ crates/bridge-proof/protocol2/Cargo.toml
@@ -13,6 +13,7 @@ rust.unused_must_use = "deny"
 rustdoc.all = "warn"
 
 [dependencies]
+strata-bridge-proof-primitives.workspace = true
 strata-crypto.workspace = true
 strata-l1tx.workspace = true
 strata-primitives.workspace = true
@@ -21,7 +22,6 @@ strata-state.workspace = true
 zkaleido.workspace = true
 zkaleido-native-adapter.workspace = true
 
-arbitrary.workspace = true
 bitcoin.workspace = true
 borsh.workspace = true
 thiserror.workspace = true
@@ -31,3 +31,4 @@ bincode.workspace = true
 serde_json.workspace = true
 hex.workspace = true
 strata-bridge-test-utils.workspace = true
+prover-test-utils.workspace = true
