+++ crates/bridge-proof/test-utils/Cargo.toml
@@ -4,6 +4,7 @@ name = "prover-test-utils"
 version = "0.1.0"
 
 [dependencies]
+strata-bridge-proof-protocol2.workspace = true
 strata-btcio.workspace = true
 strata-l1tx.workspace = true
 strata-primitives.workspace = true
@@ -11,8 +12,11 @@ strata-proofimpl-btc-blockspace.workspace = true
 strata-state.workspace = true
 zkaleido.workspace = true
 
+bincode.workspace = true
 bitcoin.workspace = true
 borsh.workspace = true
+hex.workspace = true
 serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
 sp1-core-machine = "4.0.0"
 tokio.workspace = true
